Remove board specific Mutation library modules
The only variation in collecting results for the mutation table is whether or not
boards support offsets

Boards that support offsets should set the variable MLS::Config::Mutation::OFFSET_SIZE
to the offset size supported.

Boards that do not support offsets should have this variable set to undefined or 0.

For a new board, this variable can be left undefined and the mutation script will tell you
if an offset is required. In this case, the number of requested records will be far greater
than the received amount. The variable can then be set and the offset will be used going forward.

The mutation script will also now die if the results returned for a class are missing more than
10 results from the expected count.