set a timeout for new RetsSessions to avoid waiting forever
librets by default will continue to poll for data from the server in a thread
separate from the perl process apparently. Which means even with our fancy
alarms and retries in the Rets library, it's been possible to have the perl
thread complete while some other thread is still waiting for a server response indefinitely.

This timeout should stop those queries from continuing after 15 minutes of inactivity.

If this change fails to correct the issue then
we should use another POSIX::_exit at the end of run.pl instead of a normal exit